feat: add http client parameter to pusher
Allows for specifying custom root certificates
