+++ api/internal/oci/pusher_test.go
@@ -5,200 +5,18 @@ package oci
 
 import (
 	"bytes"
-	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/json"
-	"encoding/pem"
 	"fmt"
 	"log"
-	"math/big"
+	"net/http"
 	"os"
-	"path/filepath"
 	"testing"
-	"time"
 
-	"github.com/cpuguy83/dockercfg"
 	"github.com/distribution/reference"
 	"github.com/stretchr/testify/require"
-	"github.com/stretchr/testify/suite"
-	"github.com/testcontainers/testcontainers-go"
-	"github.com/testcontainers/testcontainers-go/modules/registry"
-	"github.com/testcontainers/testcontainers-go/wait"
-	"golang.org/x/crypto/bcrypt"
+
 	loctest "sigs.k8s.io/kustomize/api/testutils/localizertest"
 	"sigs.k8s.io/kustomize/api/types"
 )
-
-type PusherTestSuite struct {
-	suite.Suite
-	registry       *registry.RegistryContainer
-	user           string
-	password       string
-	address        string
-	root_directory string
-	ca_directory   string
-	valid_auth     string
-	ctx            context.Context
-}
-
-func withOptionalRegistryCertificate(key_path string, certificate_path string) testcontainers.CustomizeRequestOption {
-	if key_path != "" && certificate_path != "" {
-		return func(req *testcontainers.GenericContainerRequest) error {
-			const container_cert_path = "/certs/registry.pem"
-			const container_key_path = "/certs/registry.key"
-			req.Files = append(req.Files,
-				testcontainers.ContainerFile{
-					HostFilePath:      certificate_path,
-					ContainerFilePath: container_cert_path,
-					FileMode:          0o644,
-				},
-				testcontainers.ContainerFile{
-					HostFilePath:      key_path,
-					ContainerFilePath: container_key_path,
-					FileMode:          0o644,
-				})
-			req.Env["REGISTRY_HTTP_TLS_CERTIFICATE"] = container_cert_path
-			req.Env["REGISTRY_HTTP_TLS_KEY"] = container_key_path
-			return nil
-		}
-	} else {
-		return func(req *testcontainers.GenericContainerRequest) error { return nil }
-	}
-}
-
-func withOptionalHtpasswd(username string, password string) testcontainers.CustomizeRequestOption {
-	if username != "" && password != "" {
-		return func(req *testcontainers.GenericContainerRequest) error {
-			if hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost); err != nil {
-				return err
-			} else {
-				entry := fmt.Sprintf("%s:%s", username, string(hash))
-				return registry.WithHtpasswd(entry)(req)
-			}
-		}
-	} else {
-		return func(req *testcontainers.GenericContainerRequest) error { return nil }
-	}
-}
-
-func createRegistry(t *testing.T, key_path string, cert_path string, username string, password string) string {
-	ctx := context.Background()
-	registryContainer, err := registry.Run(ctx, "registry:3.0.0",
-		withOptionalRegistryCertificate(key_path, cert_path),
-		withOptionalHtpasswd(username, password),
-		// Necessary to work on podman/containerized podman
-		testcontainers.WithWaitStrategy(
-			wait.ForMappedPort("5000/tcp"),
-		),
-	)
-	testcontainers.CleanupContainer(t, registryContainer)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	address, err := registryContainer.HostAddress(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return address
-}
-
-func createFileWithDirs(path string) (*os.File, error) {
-	dir := filepath.Dir(path)
-
-	if err := os.MkdirAll(dir, 0o755); err != nil {
-		return nil, err
-	}
-
-	return os.Create(path)
-}
-
-func createDockerConfig(t *testing.T, address string, user string, password string) {
-	auth_directory := t.TempDir()
-	auth_file, err := createFileWithDirs(filepath.Join(auth_directory, "config.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	auth_config, err := registry.DockerAuthConfig(address, user, password)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	encoder := json.NewEncoder(auth_file)
-	if err := encoder.Encode(dockercfg.Config{AuthConfigs: auth_config}); err != nil {
-		t.Fatal(err)
-	}
-
-	t.Setenv("DOCKER_CONFIG", auth_directory)
-}
-
-// Create a public/private key pair
-func generateSelfSignedCert(t *testing.T) (certificate string, key string) {
-	directory := t.TempDir()
-	// Generate private key
-	privateKey, err := rsa.GenerateKey(rand.Reader, 1024)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	key_path := filepath.Join(directory, "key", "registry.key")
-	key_file, err := createFileWithDirs(key_path)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if err := pem.Encode(key_file, &pem.Block{
-		Type:  "RSA PRIVATE KEY",
-		Bytes: x509.MarshalPKCS1PrivateKey(privateKey),
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	// Certificate template
-	tmpl := x509.Certificate{
-		SerialNumber: big.NewInt(1),
-		Subject: pkix.Name{
-			CommonName: "localhost",
-		},
-		NotBefore: time.Now(),
-		NotAfter:  time.Now().Add(24 * time.Hour),
-
-		KeyUsage:    x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
-		ExtKeyUsage: []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-		DNSNames: []string{
-			"localhost",
-			// So that the certificate can be used when the test process is running inside a container
-			"hub.docker.internal",
-		},
-		BasicConstraintsValid: true,
-	}
-
-	// Self-sign the certificate
-	derBytes, err := x509.CreateCertificate(rand.Reader, &tmpl, &tmpl, &privateKey.PublicKey, privateKey)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	certificate_path := filepath.Join(directory, "cert", "registry.crt")
-	certificate_file, err := createFileWithDirs(certificate_path)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if err := pem.Encode(certificate_file, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: derBytes,
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	return certificate_path, key_path
-}
 
 func AsNamedTagged(name string, tag string) reference.NamedTagged {
 	named, _ := reference.WithName(name)
@@ -443,10 +261,8 @@ func TestUntrustedCertificate(t *testing.T) {
 	username := "username"
 	password := "password"
 
-	certificate, key := generateSelfSignedCert(t)
-	// Explicitly _NOT_ setting the certificate directory
-	address := createRegistry(t, key, certificate, username, password)
-
+	// Explicitly ignoring the certificates
+	address, _ := createRegistry(t, username, password, true)
 	kustomization := map[string]string{
 		"src/kustomization.yaml": `namePrefix: test-
 `,
@@ -461,6 +277,7 @@ func TestUntrustedCertificate(t *testing.T) {
 			Namespace: "somethingnonempty",
 		},
 		targets: []reference.NamedTagged{AsNamedTagged(address+"/something", "sometag")},
+		http:    http.DefaultClient,
 	}
 
 	err := PushToOciRegistries(&pushOptions)
@@ -471,10 +288,7 @@ func TestNoCredentialFile(t *testing.T) {
 	username := "username"
 	password := "password"
 
-	certificate, key := generateSelfSignedCert(t)
-	t.Setenv("SSL_CERT_DIR", filepath.Dir(certificate))
-
-	address := createRegistry(t, key, certificate, username, password)
+	address, caCert := createRegistry(t, username, password, true)
 
 	kustomization := map[string]string{
 		"src/kustomization.yaml": `namePrefix: test-
@@ -490,6 +304,7 @@ func TestNoCredentialFile(t *testing.T) {
 			Namespace: "somethingnonempty",
 		},
 		targets: []reference.NamedTagged{AsNamedTagged(address+"/something", "sometag")},
+		http:    toClient(caCert),
 	}
 
 	err := PushToOciRegistries(&pushOptions)
@@ -497,10 +312,7 @@ func TestNoCredentialFile(t *testing.T) {
 }
 
 func TestInvalidCredentials(t *testing.T) {
-	certificate, key := generateSelfSignedCert(t)
-	t.Setenv("SSL_CERT_DIR", filepath.Dir(certificate))
-
-	address := createRegistry(t, key, certificate, "expectedusername", "expectedpassword")
+	address, caCert := createRegistry(t, "expectedusername", "expectedpassword", true)
 	createDockerConfig(t, address, "actualusername", "actualpassword")
 
 	kustomization := map[string]string{
@@ -517,6 +329,7 @@ func TestInvalidCredentials(t *testing.T) {
 			Namespace: "somethingnonempty",
 		},
 		targets: []reference.NamedTagged{AsNamedTagged(address+"/something", "sometag")},
+		http:    toClient(caCert),
 	}
 
 	err := PushToOciRegistries(&pushOptions)
@@ -527,10 +340,7 @@ func TestPush(t *testing.T) {
 	username := "username"
 	password := "password"
 
-	certificate, key := generateSelfSignedCert(t)
-	t.Setenv("SSL_CERT_DIR", filepath.Dir(certificate))
-
-	address := createRegistry(t, key, certificate, username, password)
+	address, caCert := createRegistry(t, username, password, true)
 	createDockerConfig(t, address, username, password)
 
 	kustomization := map[string]string{
@@ -547,6 +357,7 @@ func TestPush(t *testing.T) {
 			Namespace: "somethingnonempty",
 		},
 		targets: []reference.NamedTagged{AsNamedTagged(address+"/something", "sometag")},
+		http:    toClient(caCert),
 	}
 
 	err := PushToOciRegistries(&pushOptions)
